gs-app: Guarantee app reviews are in sorted order
This was already guaranteed by `GsPluginJobRefine`, but that isn’t the
right place for this code to live — and it certainly shouldn’t be
modifying the array returned by `gs_app_get_reviews()` in place.

Rework the code to be a bit better encapsulated in `GsApp`. This
introduces no functional changes.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>