gs-updates-section: Add a counter property
This exposes the number of not-yet-installed updates in this
`GsUpdatesSection`. It could be used to display a counter in the UI for
the section, or, as will happen in the following few commits, be summed
with the counters from other sections to set the overall counter (in the
shell top bar) for the number of not-yet-applied updates.

Calculating the counter in `GsUpdatesSection` makes more sense than
calculating it in `GsUpdatesPage` because the section has the
`GsAppList` containing the apps, and already has a subscription to
`app-state-changed` for that app list.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>