gs-updates-page: Calculate update count from GsUpdatesSection:counters
This avoids having to synthesise a list of all the updates from the
individual `GsAppList`s in each section, too. It should give the same
result, but a bit faster, and it allows us to use a notify signal to
update the UI when the counter changes.

The UI was already updated when the counter changes, but only via the
indirect route of the `GsPluginLoader::status-changed` signal, which can
now be removed after this commit.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>