gs-updates-page: Drop use of GsPluginLoader::status-changed signal
The signal was used to update the UI (via `gs_updates_page_update_ui_state()`),
but since the last few commits, we now ensure the UI is updated via
notifications of changes to all the relevant properties (i.e. all the values
which affect the behaviour of `gs_updates_page_update_ui_state()`), so the calls
to it in response to `GsPluginLoader::status-changed` are guaranteed to be
redundant and can be removed.

No other code in gnome-software listens to the signal, so it can be removed
entirely.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>