gs-app-query: Change :is-source to :component-kinds
This achieves two goals:
 * Makes the property more flexible, in case we want to filter by other
   kinds of component in future.
 * Moves towards renaming ‘sources’ to ‘repositories’ throughout the
   code (see #2847).

We’ve already bumped the API version this cycle, so that doesn’t need
bumping again for this API change.

This needed to introduce a new `GsComponentKindArray` boxed type so that
an array of `AsComponentKind` values could be set as a GObject property.
The boxed type is not exposed publicly in the libgnomesoftware API since
that seemed unnecessarily confusing for now. Callers can just use the
`const AsComponentKind[]` type directly.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>
Fixes: #2846