packagekit: Use async function in auto-prepare-update
Use the async version of `pk_client_get_updates()`, so that we can share
a single `GMainContext` for the thread between the call to
`pk_client_get_updates()` and the following call to
`gs_plugin_packagekit_download_async()`.

Previously, the plugin was using the sync version, which is a wrapper
provided by PackageKit. It calls the async version in its own local
`GMainContext`, so the auto-prepare-update thread was effectively
previously running two `GMainContext` iterations sequentially.

That’s bad, because it doubled the risk of a `GSource` remaining
attached to the `GMainContext` after its final iteration, and
effectively leaking.

This should help resolve memory leaks from the auto-prepare-update
thread.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>