packagekit: Run auto-prepare-update job asynchronously
Rather than running it in a worker thread. The implementation is now
entirely asynchronous, and just involves D-Bus calls to PackageKit, so
it’s safe to run in the main thread.

In particular, this fixes a leak of some internal GLib D-Bus signal
subscription allocations, which are freed on a low priority idle
callback after a `GDBusProxy` is finalised. The timing of this depends
on when the GDBus worker thread finishes its handling of the signal
subscription processing, so there’s nothing we can access from
`gs_plugin_packagekit_auto_prepare_update_thread()` which would allow us
to synchronise with that. It’s easier overall to run the whole job
asynchronously.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>