packagekit: Move a finish method to clarify things
Move the `GAsyncReadyCallback` for the whole ‘auto-prepare-update’
process to immediately below the final callback in the async callback
chain. Rename it to fit the same pattern as the other methods.

This means that when the code in
`gs_plugin_packagekit_run_prepare_update_cb()` and the callback methods
below it is read, it’s in the order it’ll be executed in.

This introduces no functional changes.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>