update-monitor: improve logic for deciding when to show notification
We are inappropriately considering only how long it has been since the
last update notification was displayed. If it has been more than a few
days since the last notification, we say updates are outdated. But
checking the last notification timestamp really only makes sense for
deciding whether we are nagging the user with too many notifications; we
shouldn't be checking it to decide whether the system is outdated or
not. The criteria for deciding that is really based on the time of the
last check for updates or the time that updates were last installed,
neither of which are examined currently.

This commit fixes the logic such that I will no longer be notified that
"Updates Are Out of Date" unless it has actually been 7 days since the
last check for update. I have automatic updates enabled, so I expect
GNOME Software to check for updates daily.

Additionally, we are supposed to nag the user daily if it has been more
than 7 days since the last check for updates. The current code nags if
it has been exactly 4 days or more than 7 days since the last nag, but
this doesn't seem to be based on the design. Fix that.

I also consistently consider can_download as part of should_download.

Fixes #2856