update-monitor: simplify check for allow-updates
When the allow-updates setting is disabled, GNOME Software is only used
for installing apps and does not have the updates page available. All
updates-related functionality is disabled.

Currently we're checking for this condition only in the one place it's
really needed. All the other branches should be unreachable because they
require the user to have performed a check for updates, which is not
possible when allow-updates is disabled. So this shouldn't actually
change anything, but it's more logical.