meson: Bump to libadwaita 1.8
This is safe to depend on it, as it’s available in GNOME 50.

It requires transitively bumping the GTK dependency too, as libadwaita
has a higher dependency than gnome-software did.

We can’t depend on any higher versions of either module, as more recent
versions of GTK start depending on a very new version of librsvg, which
is a complete pain to build in CI.