Switch to systemd-based startup
As of GNOME 49, gnome-session no longer uses .desktop files to autostart
services. Instead, we entirely rely on systemd. This means that we need
to replace service .desktop files with systemd units. For more details,
see the tracking issue[1]

gnome-software is, effectively, a session service. It starts at login,
and runs in the background to manage auto-updates. So it should be using
a systemd unit file.

This way, we gain some systemd goodies. For instance, if gnome-software
crashes it'll now be auto-restarted

[1]: https://gitlab.gnome.org/GNOME/gnome-session/-/issues/161