gs-appstream: Hide error message loading from files which do not exist
While running ./libexec/gnome-software-cmd --verbose in development it was
noticed that Gs was attempting to load more than one time from paths or
files that are not present upon this system. The issue being addressed here
is that in the g_debug log, the word Error was propagated to the output,
causing additional time for investigation.

This adjustment, places a more friendly message after an attempt is made
to load a non-existing file.