flatpak: Always use system installations
These could be skipped when the logged-in user could not update AppStream
data for the system installations, but the installation can be used even
without that, as it can show (possibly installed) apps and the user can
be asked for an admin password when installing/removing apps from it.

This change, instead of completely hiding the system installations, rather
does not update it. It can use outdated AppStream data, but it's better
than the confusion about missing system installations.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2862

Co-authored-by: Philip Withnall <philip@tecnocode.co.uk>