gs-featured-carousel: Remove unneeded code
The comment references libhandy, but here is used libadwaita. Even though
it still applies, it's not needed to carry it over. When being at it, use
provided function to get the nth page widget, instead of inventing its
own function. Also change the type of the `new_page` variable, as it
means the page index, which is not a double.