gs-update-monitor: Do not reset update notification timestamp on found historical updates
I added the reset of the update notification timestamp within changes for the https://gitlab.gnome.org/GNOME/gnome-software/-/issues/947 ,
but it was rather due to my misunderstanding of the notification about
past/historical updates and to-be-done updates. It can happen that
the user runs an offline update and then turns off the machine for
a month and when the user turns the machine again, it can both
be notified about made updates a month ago and the repositories
can have many new updates, then the user should be notified about
both things, not that the past updates reset notification timestamp
for the new/pending updates.