gs-application: Flip when to show commit ID as a version identifier
Lets have a set of build types which should show a commit identifier
as the app version and use the project version for everything else,
rather than using the project version for the "release" build type
and the commit identifier for everything else.

The "plain" build type is just like the "release" build type for
this use case.