gs-application: Always build gs-dbus-helper
Despite the name of the D-Bus interface ("org.freedesktop.PackageKit",
which is very unfortunate name), only some small parts of the D-Bus helper
use the PackageKit directly. Most of the parts call generic gnome-software
actions, which can provide useful results also when the PackageKit is
not used (think of rpm-ostree, EndlessOS, ...).

The parts which do require PackageKit return not-supported when it is not
built with the PackageKit.