gs-common: Refactor getter function for app data dirs
Refactor the getter function to return both the path to the data dir
and its state. This allows differentiation between an app data dir which
we don’t know about (e.g. for a non-sandboxed app), and an app data dir
which we know about and know to not exist.

This introduces no behavioural changes at the moment, but will allow for
UI improvements in the following commit.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>