src: Rename GsUpdateDialog to GsInstalledUpdatesDialog
This makes it a lot clearer what it’s meant to show, because being named
`GsUpdateDialog` made it seem like it should be used on `GsUpdatesPage`,
but it’s mostly not.

Things are slightly more complicated, as it’s also used (via the
secondary `gs_installed_updates_dialog_new_for_app()` constructor) to
show the details of a specific available update. A following commit will
split that out to a separate dialog class.

This introduces no functional changes, it’s just a rename of the class
and its files.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>