src: Add GsAppUpdateDetailsDialog
This will be used by `GsUpdatesSection` to show the details of a
specific app in the list of available updates. Previously, this was done
by a special mode of `GsInstalledUpdatesDialog` (née `GsUpdateDialog`),
but that was confusing.

The new dialog is a copy of `GsInstalledUpdatesDialog`, with the only
changes to the code being:
 - Renaming it
 - Dropping the `new()` constructor and renaming `new_for_app()` to
   `new()`

The rest of the code is unchanged and should not need review. The
following commit will drop dead code from both dialogs now they have
been split.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>