gs-shell: Don’t present multiple copies of dialogs if asked
There are a few dialogs which `GsShell` can show in response to a
command line argument, like `gnome-software --mode=updated` (the
installed updates dialog) or `gnome-software --prefs` (the preferences
dialog).

If you were to run one of those commands multiple times, you’d end up
with multiple instances of the dialog stacked on top of each other.

Fix that by saving a weak pointer to the dialog, and re-presenting the
existing dialog if asked to, rather than creating a new one each time.
Only create a new one if the previous one was closed.

The `GsReposDialog` isn’t triggerable from the command line (as far as I
know), but since the code for it was alongside the code for the
installed updates and preferences dialogs, it made sense to change it at
the same time too.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>