gs-page: Clarify why the user is not shown a ‘delete data’ button
When uninstalling an app, the user gets a choice about whether to delete
the app’s saved data, iff gnome-software knows where the app stores its
data (i.e. for flatpaks).

If the app hasn’t saved any data, we don’t give the user a choice
(because it would be a no-op). It could be confusing to the user as to
why some apps give them a choice and others don’t — and it probably
won’t be obvious that this is because some apps have saved data to the
home directory and others haven’t.

Clarify that by adding an extra sentence to the confirmation message.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2911